Move show and log functions out of helpers
