+++ extensions/ql-vscode/src/language-support/ast-viewer/ast-viewer.ts
@@ -24,10 +24,10 @@ import {
   isLineColumnLoc,
 } from "../../pure/bqrs-utils";
 import { DisposableObject } from "../../pure/disposable-object";
-import { showAndLogExceptionWithTelemetry } from "../../helpers";
 import { asError, getErrorMessage } from "../../pure/helpers-pure";
 import { redactableError } from "../../pure/errors";
 import { AstViewerCommands } from "../../common/commands";
+import { showAndLogExceptionWithTelemetry } from "../../common/vscode/log";
 
 export interface AstItem {
   id: BqrsId;
