+++ extensions/ql-vscode/src/query-history/store/query-history-store.ts
@@ -1,7 +1,6 @@
 import { pathExists, remove, mkdir, writeFile, readJson } from "fs-extra";
 import { dirname } from "path";
 
-import { showAndLogExceptionWithTelemetry } from "../../helpers";
 import {
   asError,
   asyncFilter,
@@ -13,6 +12,7 @@ import { redactableError } from "../../pure/errors";
 import { QueryHistoryDto, QueryHistoryItemDto } from "./query-history-dto";
 import { mapQueryHistoryToDomainModel } from "./query-history-dto-mapper";
 import { mapQueryHistoryToDto } from "./query-history-domain-mapper";
+import { showAndLogExceptionWithTelemetry } from "../../common/vscode/log";
 
 const ALLOWED_QUERY_HISTORY_VERSIONS = [1, 2];
